+++ Src/Private/Get-RequiredFeature.ps1
@@ -34,15 +34,15 @@ function Get-RequiredFeature {
         if ($OSType -eq 'WorkStation') {
             $RequiredFeature = Get-WindowsCapability -Online -Name $Name
             if ($RequiredFeature.State -ne 'Installed') {
-                throw "$Name is required to run the Microsoft AD As Built Report. Run 'Add-WindowsCapability -online -Name '$($Name)'' to install the required modules. https://github.com/AsBuiltReport/AsBuiltReport.Microsoft.AD"
+                throw "$Name is required to run the Microsoft AD As Built Report. Run 'Add-WindowsCapability -online -Name '$($Name)'' to install the required modules. Review requirements at https://github.com/AsBuiltReport/AsBuiltReport.Microsoft.AD"
             }
         } elseif ($OSType -eq 'Server' -or $OSType -eq 'DomainController') {
             $RequiredFeature = Get-WindowsFeature -Name $Name
             if ($RequiredFeature.InstallState -ne 'Installed') {
-                throw "$Name is required to run the Microsoft AD As Built Report. Run 'Install-WindowsFeature -Name '$($Name)'' to install the required modules. https://github.com/AsBuiltReport/AsBuiltReport.Microsoft.AD"
+                throw "$Name is required to run the Microsoft AD As Built Report. Run 'Install-WindowsFeature -Name '$($Name)'' to install the required modules. Review requirements at https://github.com/AsBuiltReport/AsBuiltReport.Microsoft.AD"
             }
         } else {
-            throw "Unable to validate if $Name is installed. https://github.com/AsBuiltReport/AsBuiltReport.Microsoft.AD"
+            throw "Unable to validate if $Name is installed. Review requirements at https://github.com/AsBuiltReport/AsBuiltReport.Microsoft.AD"
         }
     }
     end {}