+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add support for PowerShell 7
 - Add function to Invoke-Command with timeout support (Jobs Scheduler)
+- Add FileName parameter to Get-AbrDiagrammer to allow custom file names for diagram outputs
 
 ### Changed
 
@@ -24,11 +25,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update CodeQL action to v4
 - Migrate function located in SharedUtilsFunctions.ps1 to its own file for better modularity and maintainability
 - Change export diagrams format to "PDF" by default
+- Improve Get-RequiredFeatures function to better handle feature retrieval and error handling
 
 ### Fixed
 
-- Fix HealthCheck in Get-AbrADDomainLastBackup script to correctly evaluate the 'Last Backup in Days' property as an integer.
-- Fix 'Orphaned GPO' section displaying data when no orphaned GPOs are found in Get-AbrADGPO script.
+- Fix HealthCheck in Get-AbrADDomainLastBackup script to correctly evaluate the 'Last Backup in Days' property as an integer
+- Fix 'Orphaned GPO' section displaying data when no orphaned GPOs are found in Get-AbrADGPO script
+- Fix diagram export to allow generate Trusts diagrams per child domain
 
 ## [0.9.7] - 2025-11-14